portability lint
git-svn-id: file:///svn/unbound/trunk@2193 be551aaa-1e26-0410-a405-d3ace91eadb9

+++ util/tube.c
@@ -401,7 +401,7 @@ pollit(int fd, struct timeval* t)
 		return 0;
 	}
 	errno = 0;
-	return FD_ISSET(fd, &r);
+	return (int)(FD_ISSET(fd, &r));
 }
 
 int tube_poll(struct tube* tube)
